smart_agri_project/backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pydantic>=1.10.0
python-dotenv>=1.0.0
motor>=3.1.1          # async MongoDB driver
pymongo>=4.3.0        # optional sync access if needed
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
joblib>=1.2.0
Pillow>=10.0.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
